Split workspace configuration options into common and browser.
