+++ src/vs/workbench/contrib/url/browser/externalUriResolver.ts
@@ -7,11 +7,12 @@ import { Disposable } from 'vs/base/common/lifecycle';
 import { IOpenerService } from 'vs/platform/opener/common/opener';
 import { IWorkbenchContribution } from 'vs/workbench/common/contributions';
 import { IWorkbenchEnvironmentService } from 'vs/workbench/services/environment/common/environmentService';
+import { IBrowserWorkbenchEnvironmentService } from 'vs/workbench/services/environment/browser/environmentService';
 
 export class ExternalUriResolverContribution extends Disposable implements IWorkbenchContribution {
 	constructor(
 		@IOpenerService _openerService: IOpenerService,
-		@IWorkbenchEnvironmentService _workbenchEnvironmentService: IWorkbenchEnvironmentService,
+		@IWorkbenchEnvironmentService _workbenchEnvironmentService: IBrowserWorkbenchEnvironmentService,
 	) {
 		super();
