+++ src/vs/workbench/services/update/browser/updateService.ts
@@ -7,6 +7,7 @@ import { Event, Emitter } from 'vs/base/common/event';
 import { IUpdateService, State, UpdateType } from 'vs/platform/update/common/update';
 import { registerSingleton } from 'vs/platform/instantiation/common/extensions';
 import { IWorkbenchEnvironmentService } from 'vs/workbench/services/environment/common/environmentService';
+import { IBrowserWorkbenchEnvironmentService } from 'vs/workbench/services/environment/browser/environmentService';
 import { IHostService } from 'vs/workbench/services/host/browser/host';
 import { Disposable } from 'vs/base/common/lifecycle';
 
@@ -39,7 +40,7 @@ export class BrowserUpdateService extends Disposable implements IUpdateService {
 	}
 
 	constructor(
-		@IWorkbenchEnvironmentService private readonly environmentService: IWorkbenchEnvironmentService,
+		@IWorkbenchEnvironmentService private readonly environmentService: IBrowserWorkbenchEnvironmentService,
 		@IHostService private readonly hostService: IHostService
 	) {
 		super();
