+++ src/vs/workbench/workbench.web.api.ts
@@ -19,329 +19,8 @@ import { IWorkspaceProvider, IWorkspace } from 'vs/workbench/services/host/brows
 import { CommandsRegistry } from 'vs/platform/commands/common/commands';
 import { IProductConfiguration } from 'vs/platform/product/common/productService';
 import { mark } from 'vs/base/common/performance';
-
-interface IResourceUriProvider {
-	(uri: URI): URI;
-}
-
-interface IStaticExtension {
-	packageJSON: IExtensionManifest;
-	extensionLocation: URI;
-	isBuiltin?: boolean;
-}
-
-interface ICommonTelemetryPropertiesResolver {
-	(): { [key: string]: any };
-}
-
-interface IExternalUriResolver {
-	(uri: URI): Promise<URI>;
-}
-
-interface ITunnelProvider {
-
-	/**
-	 * Support for creating tunnels.
-	 */
-	tunnelFactory?: ITunnelFactory;
-
-	/**
-	 * Support for filtering candidate ports
-	 */
-	showPortCandidate?: IShowPortCandidate;
-}
-
-interface ITunnelFactory {
-	(tunnelOptions: ITunnelOptions): Promise<ITunnel> | undefined;
-}
-
-interface ITunnelOptions {
-	remoteAddress: { port: number, host: string };
-
-	/**
-	 * The desired local port. If this port can't be used, then another will be chosen.
-	 */
-	localAddressPort?: number;
-
-	label?: string;
-}
-
-interface ITunnel extends IDisposable {
-	remoteAddress: { port: number, host: string };
-
-	/**
-	 * The complete local address(ex. localhost:1234)
-	 */
-	localAddress: string;
-
-	/**
-	 * Implementers of Tunnel should fire onDidDispose when dispose is called.
-	 */
-	onDidDispose: Event<void>;
-}
-
-interface IShowPortCandidate {
-	(host: string, port: number, detail: string): Promise<boolean>;
-}
-
-interface ICommand {
-
-	/**
-	 * An identifier for the command. Commands can be executed from extensions
-	 * using the `vscode.commands.executeCommand` API using that command ID.
-	 */
-	id: string,
-
-	/**
-	 * A function that is being executed with any arguments passed over. The
-	 * return type will be send back to the caller.
-	 *
-	 * Note: arguments and return type should be serializable so that they can
-	 * be exchanged across processes boundaries.
-	 */
-	handler: (...args: any[]) => unknown;
-}
-
-interface IHomeIndicator {
-
-	/**
-	 * The link to open when clicking the home indicator.
-	 */
-	href: string;
-
-	/**
-	 * The icon name for the home indicator. This needs to be one of the existing
-	 * icons from our Codicon icon set. For example `sync`.
-	 */
-	icon: string;
-
-	/**
-	 * A tooltip that will appear while hovering over the home indicator.
-	 */
-	title: string;
-}
-
-interface IDefaultSideBarLayout {
-	visible?: boolean;
-	containers?: ({
-		id: 'explorer' | 'run' | 'scm' | 'search' | 'extensions' | 'remote' | string;
-		active: true;
-		order?: number;
-		views?: {
-			id: string;
-			order?: number;
-			visible?: boolean;
-			collapsed?: boolean;
-		}[];
-	} | {
-		id: 'explorer' | 'run' | 'scm' | 'search' | 'extensions' | 'remote' | string;
-		active?: false;
-		order?: number;
-		visible?: boolean;
-		views?: {
-			id: string;
-			order?: number;
-			visible?: boolean;
-			collapsed?: boolean;
-		}[];
-	})[];
-}
-
-interface IDefaultPanelLayout {
-	visible?: boolean;
-	containers?: ({
-		id: 'terminal' | 'debug' | 'problems' | 'output' | 'comments' | string;
-		order?: number;
-		active: true;
-	} | {
-		id: 'terminal' | 'debug' | 'problems' | 'output' | 'comments' | string;
-		order?: number;
-		active?: false;
-		visible?: boolean;
-	})[];
-}
-
-interface IDefaultView {
-	readonly id: string;
-}
-
-interface IDefaultEditor {
-	readonly uri: UriComponents;
-	readonly openOnlyIfExists?: boolean;
-	readonly openWith?: string;
-}
-
-interface IDefaultLayout {
-	/** @deprecated Use views instead (TODO@eamodio remove eventually) */
-	readonly sidebar?: IDefaultSideBarLayout;
-	/** @deprecated Use views instead (TODO@eamodio remove eventually) */
-	readonly panel?: IDefaultPanelLayout;
-	readonly views?: IDefaultView[];
-	readonly editors?: IDefaultEditor[];
-}
-
-interface IProductQualityChangeHandler {
-
-	/**
-	 * Handler is being called when the user wants to switch between
-	 * `insider` or `stable` product qualities.
-	 */
-	(newQuality: 'insider' | 'stable'): void;
-}
-
-interface IWorkbenchConstructionOptions {
-
-	//#region Connection related configuration
-
-	/**
-	 * The remote authority is the IP:PORT from where the workbench is served
-	 * from. It is for example being used for the websocket connections as address.
-	 */
-	readonly remoteAuthority?: string;
-
-	/**
-	 * The connection token to send to the server.
-	 */
-	readonly connectionToken?: string;
-
-	/**
-	 * An endpoint to serve iframe content ("webview") from. This is required
-	 * to provide full security isolation from the workbench host.
-	 */
-	readonly webviewEndpoint?: string;
-
-	/**
-	 * A factory for web sockets.
-	 */
-	readonly webSocketFactory?: IWebSocketFactory;
-
-	/**
-	 * A provider for resource URIs.
-	 */
-	readonly resourceUriProvider?: IResourceUriProvider;
-
-	/**
-	 * Resolves an external uri before it is opened.
-	 */
-	readonly resolveExternalUri?: IExternalUriResolver;
-
-	/**
-	 * A provider for supplying tunneling functionality,
-	 * such as creating tunnels and showing candidate ports to forward.
-	 */
-	readonly tunnelProvider?: ITunnelProvider;
-
-	//#endregion
-
-
-	//#region Workbench configuration
-
-	/**
-	 * A handler for opening workspaces and providing the initial workspace.
-	 */
-	readonly workspaceProvider?: IWorkspaceProvider;
-
-	/**
-	 * The user data provider is used to handle user specific application
-	 * state like settings, keybindings, UI state (e.g. opened editors) and snippets.
-	 */
-	userDataProvider?: IFileSystemProvider;
-
-	/**
-	 * Session id of the current authenticated user
-	 */
-	readonly authenticationSessionId?: string;
-
-	/**
-	 * Enables user data sync by default and syncs into the current authenticated user account using the provided [authenticationSessionId}(#authenticationSessionId).
-	 */
-	readonly enableSyncByDefault?: boolean;
-
-	/**
-	 * The credentials provider to store and retrieve secrets.
-	 */
-	readonly credentialsProvider?: ICredentialsProvider;
-
-	/**
-	 * Add static extensions that cannot be uninstalled but only be disabled.
-	 */
-	readonly staticExtensions?: ReadonlyArray<IStaticExtension>;
-
-	/**
-	 * Service end-point hosting builtin extensions
-	 */
-	readonly builtinExtensionsServiceUrl?: string;
-
-	/**
-	 * Support for URL callbacks.
-	 */
-	readonly urlCallbackProvider?: IURLCallbackProvider;
-
-	/**
-	 * Support adding additional properties to telemetry.
-	 */
-	readonly resolveCommonTelemetryProperties?: ICommonTelemetryPropertiesResolver;
-
-	/**
-	 * A set of optional commands that should be registered with the commands
-	 * registry.
-	 *
-	 * Note: commands can be called from extensions if the identifier is known!
-	 */
-	readonly commands?: readonly ICommand[];
-
-	/**
-	 * Optional default layout to apply on first time the workspace is opened.
-	 */
-	readonly defaultLayout?: IDefaultLayout;
-
-	//#endregion
-
-
-	//#region Update/Quality related
-
-	/**
-	 * Support for update reporting
-	 */
-	readonly updateProvider?: IUpdateProvider;
-
-	/**
-	 * Support for product quality switching
-	 */
-	readonly productQualityChangeHandler?: IProductQualityChangeHandler;
-
-	//#endregion
-
-
-	//#region Branding
-
-	/**
-	 * Optional home indicator to appear above the hamburger menu in the activity bar.
-	 */
-	readonly homeIndicator?: IHomeIndicator;
-
-	/**
-	 * Optional override for the product configuration properties.
-	 */
-	readonly productConfiguration?: Partial<IProductConfiguration>;
-
-	//#endregion
-
-
-	//#region Diagnostics
-
-	/**
-	 * Current logging level. Default is `LogLevel.Info`.
-	 */
-	readonly logLevel?: LogLevel;
-
-	/**
-	 * Whether to enable the smoke test driver.
-	 */
-	readonly driver?: boolean;
-
-	//#endregion
-}
+import { IStaticExtension, ITunnelProvider, ITunnelFactory, ITunnel, ITunnelOptions, IShowPortCandidate } from 'vs/workbench/common/options';
+import { IBrowserWorkbenchConstructionOptions, IResourceUriProvider, IProductQualityChangeHandler, ICommonTelemetryPropertiesResolver, IExternalUriResolver, ICommand, IHomeIndicator, IDefaultView, IDefaultEditor, IDefaultLayout, IDefaultPanelLayout, IDefaultSideBarLayout } from 'vs/workbench/browser/options';
 
 interface IWorkbench {
 	commands: {
@@ -358,7 +37,7 @@ interface IWorkbench {
 let created = false;
 let workbenchPromiseResolve: Function;
 const workbenchPromise = new Promise<IWorkbench>(resolve => workbenchPromiseResolve = resolve);
-async function create(domElement: HTMLElement, options: IWorkbenchConstructionOptions): Promise<void> {
+async function create(domElement: HTMLElement, options: IBrowserWorkbenchConstructionOptions): Promise<void> {
 
 	// Mark start of workbench
 	mark('didLoadWorkbenchMain');
@@ -411,7 +90,7 @@ export {
 
 	// Factory
 	create,
-	IWorkbenchConstructionOptions,
+	IBrowserWorkbenchConstructionOptions,
 	IWorkbench,
 
 	// Basic Types
